Replace the ui.
