+++ unvesting-app/index.html
@@ -31,6 +31,6 @@
 	>
     </script>
 </head>
-	<body id="app" class="container-fluid home-bluer">
+	<body id="app" class="">
 	</body>
 </html>
