+++ unvesting-app/src/components/home/Navbar.svelte
@@ -1,0 +1,11 @@
+<section>
+  <nav class="navbar bg-nav-primary">
+    <div class="container-fluid">
+      <span class="navbar-brand mb-0 h1">
+        <a href="/">
+          <img class="logo-size" src="assets/tf_unVisting.png" alt="logo" />
+        </a>
+      </span>
+    </div>
+  </nav>
+</section>
